requirements.txt
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
attrdict==2.0.1
Babel==2.15.0
bce-python-sdk==0.9.17
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.2
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.53.0
fsspec==2024.6.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imgaug==0.4.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.0
paddlepaddle==2.6.1
pandas==2.2.2
parso==0.8.4
pdf2docx==0.5.8
pillow==10.4.0
platformdirs==4.2.2
premailer==3.10.0
prompt_toolkit==3.0.47
protobuf==3.20.2
psutil==6.0.0
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycryptodome==3.20.0
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.3
rarfile==4.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12.1
tbb==2021.13.0
termcolor==2.4.0
tifffile==2024.6.18
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.0.3
